+++ layout/pre-view/pre-view.config.js
@@ -1,0 +1,14 @@
+module.exports = {
+  folder: './views',
+  port: 8080,
+  hostname: '127.0.0.1',
+  open: true,
+  binary: null,
+  ini: null,
+  hotkeys: {
+    selector: 'alt+enter',
+    states: 'alt+shift'
+  },
+  js: [],
+  css: []
+}