+++ layout/pre-view/src/php/cache/1db80d6e3950ad67612436a3aa97c77f551b60a3.php
@@ -1,1 +1,0 @@
-<?php /**PATH /Users/olivierbossel/data/web/coffeekraken/coffeekraken/layout/pre-view/views/atoms/buttons.blade.php ENDPATH**/ ?>