+++ layout/pre-view/src/php/views/iframe.blade.php
@@ -1,0 +1,12 @@
+<html>
+  <head>
+    <link rel="stylesheet" href="{{ $css }}" type="text/css">
+  </head>
+  <body>
+
+    {!! $html !!}
+
+    <script src="{{ $js }}"></script>
+    <script src="/dist/js/iframe.bundle.js"></script>
+  </body>
+</html>