+++ layout/starter-kit/src/js/app.js
@@ -1,5 +1,0 @@
-/**
- * Polyfills
- */
-import "./polyfills/fastclick";
-import "./polyfills/cssuseragent";