+++ layout/starter-kit/views/atoms/button/button.blade.php
@@ -1,0 +1,3 @@
+<a href="{{ $href }}" name="{{ $name }}" class="btn">
+  {{ $text }}
+</a>