+++ webcomponent/select-webcomponent/CHANGELOG.md
@@ -5,4 +5,10 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [1.0.0] - 2019-11-26
+## [0.0.3] - 2019-12-10
+
+- Make the property "box-sizing: border-box" important on the search field
+
+## [0.0.2] - 2019-12-10
+
+- Adding the property "box-sizing: border-box" on the search field